smart-chef-api/requirements.txt
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
cffi==1.15.0
cryptography==36.0.2
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
drf-spectacular==0.22.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
inflection==0.5.1
jsonschema==4.4.0
Markdown==3.3.6
mysqlclient==2.1.0
Pillow==9.1.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
schedule==1.1.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
zipp==3.7.0
